Moving away from sqlite...
Everything is really messed up, but can get ingame.. character models
are not loading tho.

+++ Game/Constants/Character/Gender.cs
@@ -1,6 +1,6 @@
 ﻿namespace CobolWow.Game.Constants.Character
 {
-   public enum Gender : byte
+   public enum Gender : int
    {
       Male = 0,
       Female = 1
